+++ utils/test-utils/utilities/ClassLoaderTestUtils.ts
@@ -1,3 +1,5 @@
+import { pwd } from "shelljs";
+
 //  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
 //
 //   Copyright 2016-2017 Pascal ECHEMANN.
@@ -19,6 +21,7 @@
  */
 
 // Utilities:
-export const VALID_CLASS:string = "../../../../../dist/com/jec/commons/files/reflect/impl/BasicDecoratorProperties";
+const PWD:string = pwd();
+export const VALID_CLASS:string = PWD + "/dist/com/jec/commons/files/reflect/impl/BasicDecoratorProperties";
 export const INVALID_CLASS_PATH:string = "invalid/path/to/module";
-export const INVALID_CLASS_FILE:string = "../../../../../utils/test-utils/classes/invalid-class-module";
+export const INVALID_CLASS_FILE:string = PWD + "/utils/test-utils/classes/invalid-class-module";