+++ utils/test-utils/utilities/StaticResourceTestUtils.ts
@@ -15,6 +15,7 @@
 //   limitations under the License.
 
 import {ClassLoader} from "../../../src/com/jec/commons/lang/ClassLoader";
+import {DefaultClassLoader} from "../../../src/com/jec/commons/lang/utils/DefaultClassLoader";
 import {DecoratorConnectorManager} from "../../../src/com/jec/commons/jcad/spi/DecoratorConnectorManager";
 import {JcadContextManager} from "../../../src/com/jec/commons/jcad/spi/JcadContextManager";
 import {AbstractDecoratorConnector} from "../../../src/com/jec/commons/jcad/spi/AbstractDecoratorConnector";
@@ -31,7 +32,7 @@ import {ConfigConnector} from "../classes/ConfigConnector";
 */
 
 // Utilities:
-const LOADER:ClassLoader = new ClassLoader();
+const LOADER:ClassLoader = new DefaultClassLoader();
 const VALID_CLASS:string = process.cwd() + "/utils/test-utils/classes/StaticResourceTestClass";
 export const URL_PATTERN:string = "/url/pattern/*";
 export const CACHE_CONTROL_PARAMS:CacheControlPolicy = {